MingGW: emulate pipe(2)
The pipe(2) function is not available on Windows,
in favour of a broader _pipe() call.

Wrap it in a macro to maintain the interface

Solves error:
```
src/DiskIO/DiskThreads/CommIO.cc: In static member function 'static void CommIO::Initialize()':
src/DiskIO/DiskThreads/CommIO.cc:26:9: error: 'pipe' was not declared in this scope; did you mean '_pipe'?
   26 |     if (pipe(DonePipe)) {}
      |         ^~~~
      |         _pipe
```